refactor(inspector): extract params panel
